+++ src/index.ts
@@ -160,6 +160,35 @@ export default class WidthPlugin extends Plugin {
 
         window.addEventListener('beforeunload', this.beforeUnloadBindThis);
 
+        this.addCommand({
+            langKey: 'plugin-width.plus',
+            langText: 'Make editor wider',
+            hotkey: '⌥=',
+            callback: () => {
+                let width = this.settingUtils.get('width');
+                if (width + 5 <= 100) {
+                    let width = this.settingUtils.get('width');
+                    width += 5;
+                    this.settingUtils.set('width', width);
+                    document.documentElement.style.setProperty('--centerWidth', `${width}%`);
+                }
+            }
+        });
+        this.addCommand({
+            langKey: 'plugin-width.minus',
+            langText: 'Make editor narrower',
+            hotkey: '⌥-',
+            callback: () => {
+                let width = this.settingUtils.get('width');
+                if (width - 5 >= 40) {
+                    let width = this.settingUtils.get('width');
+                    width -= 5;
+                    this.settingUtils.set('width', width);
+                    document.documentElement.style.setProperty('--centerWidth', `${width}%`);
+                }
+            }
+        });
+
         changelog(
             this, 'i18n/changelog.md'
         ).then(({ Dialog }) => {
@@ -283,6 +312,13 @@ export default class WidthPlugin extends Plugin {
             type: 'checkbox',
             title: this.i18n.setting.enableMobile.title,
             description: this.i18n.setting.enableMobile.description,
+        });
+        this.settingUtils.addItem({
+            key: 'enableHotkey',
+            value: true,
+            type: 'checkbox',
+            title: this.i18n.setting.enableHotkey.title,
+            description: this.i18n.setting.enableHotkey.description,
         });
         this.settingUtils.addItem({
             key: 'mode',