+++ src/i18n/changelog.md
@@ -2,9 +2,11 @@
 
 - fix: 修复样式错误 [#19](https://github.com/frostime/siyuan-center-width/issues/19)
 - feat: 适配 Preview 模式 [#15](https://github.com/frostime/siyuan-center-width/issues/15)
+- feat: 使用快捷键调整宽度 [#17](https://github.com/frostime/siyuan-center-width/issues/17)
 
 ---
 
 - fix: Style error [#19](https://github.com/frostime/siyuan-center-width/issues/19)
 - feat: Adaptation to Preview Mode [#15](https://github.com/frostime/siyuan-center-width/issues/15)
+- feat: Using hotkey to adjust width [#17](https://github.com/frostime/siyuan-center-width/issues/17)